Upgrades & Updated to Libuv to most recent release
Upgraded current Socketpair checks for winloop to have faster checks in place and I am currently working on a way to optimize more features if they can be done or implemented. But feel free to write an issue if or pull request if you have any new ideas or optimizations that you want to see added or implemented, test suite coming soon.